MyFunctionProj/requirements.txt
azure-functions
azure-core==1.1.1
azure-cosmos==4.0.0b6
certifi==2019.11.28
chardet==3.0.4
idna==2.8
requests==2.22.0
six==1.13.0
urllib3==1.25.7